Implement gun system

+++ RDC/include/Enemy.hpp
@@ -4,12 +4,12 @@
 
 class Enemy : public RenderObject {
 protected:
-	int hp{100};
+	float hp{100};
 	Vector2f pos;
 
 public:
 	void move(Vector2f offset) override;
 	void move(Vector2f offset, float dt);
-	int getHp();
-	void takeDamage(int damage);
+	float getHp();
+	void takeDamage(float damage);
 };
